+++ src/Mibo.Raylib/Next/Graphics2D/ParticleDraw.fs
@@ -1,14 +1,23 @@
 namespace Mibo.Elmish.Next.Graphics2D
 
-open Mibo.Elmish.Next.Graphics2D.Base
-
+open System.Buffers
 open Raylib_cs
 open Mibo.Elmish.Next
+open Mibo.Elmish.Next.Graphics2D.Base
 open Mibo.Elmish.Graphics2D.Lighting
 
 module ParticleDraw =
 
-  let inline particles
+  let private pool = ArrayPool<ParticleData>.Create()
+  let private rented = ResizeArray<ParticleData[]>()
+
+  let returnAll() =
+    for arr in rented do
+      pool.Return(arr, false)
+
+    rented.Clear()
+
+  let particles
     (texture: Texture2D)
     (particles: Particle2D[])
     (count: int)
@@ -19,7 +28,8 @@ module ParticleDraw =
 
     let data: ParticleData[] =
       if count > 0 then
-        let arr = Array.zeroCreate count
+        let arr = pool.Rent count
+        rented.Add arr
 
         for i = 0 to count - 1 do
           let p = particles[i]